refactor(api): move ingest from app.py to routes/ingest.py
