+++ src/mnemolet/api/app.py
@@ -5,25 +5,20 @@ from typing import Any
 from fastapi import (
     APIRouter,
     FastAPI,
-    File,
     HTTPException,
-    Query,
-    UploadFile,
 )
 from fastapi.responses import StreamingResponse
 
 from mnemolet.api.routes.chat import api_router as chat_router
+from mnemolet.api.routes.ingest import api_router as ingest_router
 from mnemolet.config import (
-    BATCH_SIZE,
     EMBED_MODEL,
     MIN_SCORE,
     OLLAMA_MODEL,
     OLLAMA_URL,
     QDRANT_COLLECTION,
     QDRANT_URL,
-    SIZE_CHARS,
     TOP_K,
-    UPLOAD_DIR,
 )
 from mnemolet.cuore.utils.qdrant import QdrantManager
 
@@ -33,52 +28,7 @@ app = FastAPI(title="MnemoLet API", version="0.0.1")
 api_router = APIRouter()
 
 api_router.include_router(chat_router)
-
-
-@api_router.post("/ingest")
-async def ingest_files(
-    files: list[UploadFile] = File(...),
-    force: bool = Query(
-        False, description="Recreate Qdrant collection before ingestion"
-    ),
-):
-    """
-    Ingest multiple files into Qdrant.
-    """
-    saved_files, result = await do_ingestion(files, force)
-
-    return {
-        "status": "ok",
-        "uploaded": saved_files,
-        "force": force,
-        "message": "Ingestion complete",
-        "ingestion": {
-            "files": result["files"],
-            "chunks": result["chunks"],
-            "time": result["time"],
-        },
-    }
-
-
-async def do_ingestion(files, force: bool = False):
-    from mnemolet.cuore.ingestion.ingest import ingest
-
-    saved_files = []
-
-    for f in files:
-        dest = UPLOAD_DIR / f.filename
-
-        content = await f.read()
-        dest.write_bytes(content)
-
-        saved_files.append(str(dest))
-
-    batch_size = BATCH_SIZE
-    result = ingest(
-        UPLOAD_DIR, batch_size, QDRANT_URL, QDRANT_COLLECTION, SIZE_CHARS, force=force
-    )
-
-    return saved_files, result
+api_router.include_router(ingest_router)
 
 
 @api_router.get("/search")
